lib/libc/aarch64/string: add strcspn optimized implementation
This is a port of the Scalar optimized variant of strcspn for amd64
to aarch64 It utilizes a LUT to speed up the function, a SIMD
variant is still under development.

Performance benchmarks are as usual generated by strperf.

See the DR for benchmark results.

Tested by:	exprun
Approved by:	fuz
Reviewed by:	fuz, emaste
Sponsored by:	Google LLC (GSoC 2024)
Differential Revision: https://reviews.freebsd.org/D46398